requirements.txt
aiohttp==3.7.4.post0
aioredis==2.0.0
amqp==5.0.7
anyio==3.4.0
asgi-lifespan==1.0.1
asgiref==3.4.1
async-timeout==3.0.1
asynctest==0.13.0
attrs==21.2.0
beanie==1.8.6
bidict==0.21.4
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dnspython==1.16.0
eventlet==0.30.2
fastapi==0.70.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httptools==0.2.0
httpx==0.18.2
idna==3.2
iniconfig==1.1.1
kombu==5.2.2
motor==2.5.1
multidict==5.1.0
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.24
py==1.11.0
pydantic==1.8.2
pymongo==3.12.1
pyparsing==3.0.6
pytest==6.2.5
pytest-asyncio==0.16.0
python-dotenv==0.19.2
python-engineio==4.3.0
python-socketio==5.5.0
pytz==2021.3
PyYAML==6.0
redis==4.0.2
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.16.0
vine==5.0.0
wcwidth==0.2.5
websockets==10.1
wrapt==1.13.3
yarl==1.7.2
pytest-cov==3.0.0
black==21.12b0
isort==5.10.1
flake8==4.0.1
PyJWT==2.3.0
inflection==0.5.1
pre-commit==2.20.0
